app/requirements.txt
# requirements.txt
# NGO Food Donation App - Complete Dependencies

# Core Flask
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.0

# Google Maps API
googlemaps==4.10.0

# Firebase
firebase-admin==6.2.0

# Database & ORM (if using SQL)
SQLAlchemy==2.0.0
psycopg2-binary==2.9.0

# Data Validation & Serialization
marshmallow==3.20.0
pydantic==2.0.0

# Environment Variables
python-dotenv==1.0.0

# Authentication (Optional)
PyJWT==2.8.0
flask-jwt-extended==4.5.0

# API Rate Limiting
Flask-Limiter==3.5.0

# Caching (Optional)
Flask-Caching==2.0.0

# Async Support
aiohttp==3.9.0

# Task Queue (Optional)
celery==5.3.0
redis==5.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.0
flake8==6.1.0
pylint==3.0.0

# Production Server
gunicorn==21.2.0

# Monitoring & Logging
python-json-logger==2.0.0

# Security
cryptography==41.0.0
